Made the app more modular
